fix complier errors

+++ Agora-C_Sharp-RTC-SDK/agorartc/src/deviceManager/AgoraRtcVideoDeviceManager.cs
@@ -59,7 +59,7 @@ namespace agora.rtc
             string jsonParam = JsonMapper.ToJson(param);
             var ret = AgoraRtcNative.CallIrisApi(_irisApiEngine,
                 AgoraApiType.FUNC_VIDEODEVICEMANAGER_SETDEVICE,
-                jsonParam, jsonParam.Length, null, 0, out _result);
+                jsonParam, (UInt64)jsonParam.Length, null, 0, out _result);
             return ret != 0 ? ret : (int) AgoraJson.GetData<int>(_result.Result, "result");
         }
 
@@ -80,7 +80,7 @@ namespace agora.rtc
             string jsonParam = JsonMapper.ToJson(param);
             var ret = AgoraRtcNative.CallIrisApi(_irisApiEngine,
                 AgoraApiType.FUNC_VIDEODEVICEMANAGER_STARTDEVICETEST,
-                jsonParam, jsonParam.Length, null, 0, out _result);
+                jsonParam, (UInt64)jsonParam.Length, null, 0, out _result);
             return ret != 0 ? ret : (int) AgoraJson.GetData<int>(_result.Result, "result");
         }
